Move txn_builder to ledger_state, rename AtomicStorage
