+++ src/ledger_state.rs
@@ -7,62 +7,325 @@
 
 //! The ledger state module.
 //!
-//! This module defines [LedgerState] and [LedgerStateStore], which provide interfaces for building
+//! This module defines [LedgerState] and [LedgerStates], which provide interfaces for building
 //! transactions and control the ledger state resource.
 
 use crate::{
     events::{EventIndex, EventSource, LedgerEvent},
-    input_records,
     key_scan::{receive_history_entry, BackgroundKeyScan},
-    records::Records,
+    records::{Record, Records},
     sparse_merkle_tree::SparseMerkleTree,
     transactions::{SignedMemos, Transaction, TransactionParams, Transactions},
-    txn_builder::{bigint_to_u256, TransactionStatus, TransferSpec},
-    Captures, CommittedTxn, CryptoSnafu, EncryptingResourceAdapter, EventSummary, KeystoreBackend,
-    KeystoreError, KeystoreModel, MintInfo, RecordAmount, TransactionError, TransactionUID,
+    Captures, EncryptingResourceAdapter, EventSummary, KeystoreBackend, KeystoreError,
+    KeystoreModel, MintInfo,
 };
+use arbitrary::{Arbitrary, Unstructured};
+use ark_serialize::*;
 use ark_std::future::Future;
 use atomic_store::{AtomicStoreLoader, RollingLog};
 use chrono::Local;
+use derive_more::*;
+use espresso_macros::ser_test;
 use futures::stream::Stream;
 use jf_cap::{
+    errors::TxnApiError,
     freeze::{FreezeNote, FreezeNoteInput},
-    keys::{FreezerKeyPair, UserAddress, UserKeyPair, UserPubKey, ViewerKeyPair},
+    keys::{
+        FreezerKeyPair, FreezerPubKey, UserAddress, UserKeyPair, UserPubKey, ViewerKeyPair,
+        ViewerPubKey,
+    },
     mint::MintNote,
     proof::freeze::FreezeProvingKey,
     proof::{mint::MintProvingKey, transfer::TransferProvingKey},
     sign_receiver_memos,
     structs::{
-        AssetCode, AssetCodeSeed, AssetDefinition, AssetPolicy, FeeInput, FreezeFlag, ReceiverMemo,
-        RecordCommitment, RecordOpening, TxnFeeInfo,
+        Amount, AssetCode, AssetCodeSeed, AssetDefinition, AssetPolicy, FeeInput, FreezeFlag,
+        ReceiverMemo, RecordCommitment, RecordOpening, TxnFeeInfo,
     },
     transfer::{TransferNote, TransferNoteInput},
     AccMemberWitness, KeyPair, MerkleLeafProof, MerklePath, Signature, TransactionNote,
 };
 use jf_primitives::aead::EncKey;
+use jf_utils::tagged_blob;
 use key_set::{KeySet, OrderByOutputs, ProverKeySet};
 use num_bigint::{BigInt, Sign};
-use num_traits::identities::Zero;
+use num_traits::identities::{One, Zero};
 use primitive_types::U256;
 use rand_chacha::ChaChaRng;
+#[cfg(test)]
+use reef::cap;
 use reef::{
     traits::{
-        Block as _, NullifierSet as _, Transaction as _, TransactionKind as _, ValidationError,
-        Validator as _,
+        Block as _, Ledger, NullifierSet as _, Transaction as _, TransactionKind as _,
+        ValidationError, Validator as _,
     },
     types::*,
-    TransactionKind, Validator, *,
+    TransactionKind, Validator,
 };
 use serde::{de::DeserializeOwned, Deserialize, Serialize};
-use snafu::ResultExt;
+use snafu::{ResultExt, Snafu};
 use std::collections::HashMap;
-use std::convert::TryInto;
+use std::convert::{TryFrom, TryInto};
+use std::fmt::Debug;
+use std::hash::{Hash, Hasher};
 use std::iter::repeat;
+use std::ops::Mul;
 use std::sync::Arc;
 
-/// A never expired target.
-pub const UNEXPIRED_VALID_UNTIL: u64 = 2u64.pow(jf_cap::constants::MAX_TIMESTAMP_LEN as u32) - 1;
+// A never expired target.
+const UNEXPIRED_VALID_UNTIL: u64 = 2u64.pow(jf_cap::constants::MAX_TIMESTAMP_LEN as u32) - 1;
 const ATOMIC_STORE_RETAINED_ENTRIES: u32 = 5;
+
+/// Record amount.
+#[derive(
+    Debug,
+    Clone,
+    Copy,
+    PartialEq,
+    Default,
+    Hash,
+    Eq,
+    From,
+    Into,
+    Add,
+    AddAssign,
+    Sub,
+    SubAssign,
+    MulAssign,
+    Div,
+    DivAssign,
+    Rem,
+    RemAssign,
+    PartialOrd,
+    Ord,
+    Sum,
+    Deserialize,
+    Serialize,
+    Display,
+    FromStr,
+    LowerHex,
+)]
+#[serde(from = "u128", into = "u128")]
+#[from(types(u128, u64, u32, u8))]
+#[into(types(u128))]
+pub struct RecordAmount(pub Amount);
+
+impl RecordAmount {
+    pub fn max() -> Self {
+        (2u128.pow(127) - 1).into()
+    }
+
+    pub fn is_positive(&self) -> bool {
+        !self.is_zero()
+    }
+}
+
+impl Zero for RecordAmount {
+    fn zero() -> Self {
+        0u128.into()
+    }
+
+    fn is_zero(&self) -> bool {
+        *self == Self::zero()
+    }
+}
+
+impl One for RecordAmount {
+    fn one() -> Self {
+        1u128.into()
+    }
+}
+
+impl<T: Into<RecordAmount>> Mul<T> for RecordAmount {
+    type Output = Self;
+
+    fn mul(self, other: T) -> Self {
+        (u128::from(self) * u128::from(other.into())).into()
+    }
+}
+
+impl From<RecordAmount> for U256 {
+    fn from(amt: RecordAmount) -> U256 {
+        u128::from(amt).into()
+    }
+}
+
+/// The error that happens when converting a record amount.
+#[derive(Clone, Copy, Debug)]
+pub enum ConvertRecordAmountError {
+    OutOfRange,
+}
+
+impl TryFrom<U256> for RecordAmount {
+    type Error = ConvertRecordAmountError;
+
+    fn try_from(u: U256) -> Result<Self, Self::Error> {
+        if u <= Self::max().into() {
+            Ok(u.as_u128().into())
+        } else {
+            Err(ConvertRecordAmountError::OutOfRange)
+        }
+    }
+}
+
+impl TryFrom<BigInt> for RecordAmount {
+    type Error = ConvertRecordAmountError;
+
+    fn try_from(i: BigInt) -> Result<Self, Self::Error> {
+        if i >= 0u64.into() {
+            bigint_to_u256(i).try_into()
+        } else {
+            Err(ConvertRecordAmountError::OutOfRange)
+        }
+    }
+}
+
+// This is to make numeric literal inference work.
+//
+// When an integer literal without a suffix is provided, it defaults to `i32`. Where an
+// Into<RecordAmount> is required, such as in the wallet's public transaction building interface,
+// this trait impl can be used to do the conversion.
+impl From<i32> for RecordAmount {
+    fn from(i: i32) -> Self {
+        assert!(i >= 0);
+        (i as u128).into()
+    }
+}
+
+/// Errors that happens during a transaction.
+#[derive(Debug, Snafu)]
+#[snafu(visibility(pub))]
+pub enum TransactionError {
+    InsufficientBalance {
+        asset: AssetCode,
+        required: U256,
+        actual: U256,
+    },
+    Fragmentation {
+        asset: AssetCode,
+        amount: U256,
+        suggested_amount: U256,
+        max_records: usize,
+    },
+    TooManyOutputs {
+        asset: AssetCode,
+        max_records: usize,
+        num_receivers: usize,
+        num_change_records: usize,
+    },
+    InvalidSize {
+        asset: AssetCode,
+        num_inputs_required: usize,
+        num_inputs_actual: usize,
+        num_outputs_required: usize,
+        num_outputs_actual: usize,
+    },
+    NoFitKey {
+        num_inputs: usize,
+        num_outputs: usize,
+    },
+    CryptoError {
+        source: TxnApiError,
+    },
+    InvalidViewerKey {
+        my_key: ViewerPubKey,
+        asset_key: ViewerPubKey,
+    },
+    InvalidFreezerKey {
+        my_key: FreezerPubKey,
+        asset_key: FreezerPubKey,
+    },
+}
+
+/// Find records which can be the input to a transaction, matching the given parameters.
+fn input_records<'a, L: Ledger + 'a>(
+    records: &'a Records,
+    asset: &'a AssetCode,
+    owner: &'a UserAddress,
+    frozen: FreezeFlag,
+    now: u64,
+) -> Option<impl Iterator<Item = Record> + 'a> {
+    let spendable = records.get_spendable::<L>(&*asset, owner, frozen);
+    spendable.map(|r| r.into_iter().filter(move |record| !record.on_hold(now)))
+}
+
+/// The status of a transaction.
+#[derive(Clone, Copy, Debug, PartialEq, Eq, Serialize, Deserialize)]
+pub enum TransactionStatus {
+    Pending,
+    AwaitingMemos,
+    Retired,
+    Rejected,
+    Unknown,
+}
+
+impl std::fmt::Display for TransactionStatus {
+    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
+        match self {
+            Self::Pending => write!(f, "pending"),
+            Self::AwaitingMemos => write!(f, "accepted, waiting for owner memos"),
+            Self::Retired => write!(f, "accepted"),
+            Self::Rejected => write!(f, "rejected"),
+            Self::Unknown => write!(f, "unknown"),
+        }
+    }
+}
+
+impl TransactionStatus {
+    pub fn is_final(&self) -> bool {
+        matches!(self, Self::Retired | Self::Rejected)
+    }
+
+    pub fn succeeded(&self) -> bool {
+        matches!(self, Self::Retired)
+    }
+}
+
+/// The UID of a transaction.
+#[ser_test(arbitrary, types(cap::Ledger))]
+#[tagged_blob("TXUID")]
+#[derive(Clone, Debug, CanonicalSerialize, CanonicalDeserialize)]
+pub struct TransactionUID<L: Ledger>(pub TransactionHash<L>);
+
+impl<L: Ledger> PartialEq<TransactionUID<L>> for TransactionUID<L> {
+    fn eq(&self, other: &Self) -> bool {
+        self.0 == other.0
+    }
+}
+
+impl<L: Ledger> Eq for TransactionUID<L> {}
+
+impl<L: Ledger> Hash for TransactionUID<L> {
+    fn hash<H: Hasher>(&self, state: &mut H) {
+        Hash::hash(&self.0, state)
+    }
+}
+
+impl<'a, L: Ledger> Arbitrary<'a> for TransactionUID<L>
+where
+    TransactionHash<L>: Arbitrary<'a>,
+{
+    fn arbitrary(u: &mut Unstructured<'a>) -> arbitrary::Result<Self> {
+        Ok(Self(u.arbitrary()?))
+    }
+}
+
+/// Specifications for building a transfer.
+pub struct TransferSpec<'a> {
+    /// List of key_pairs that will be used to find the records for the transfer.
+    ///
+    /// The list may contain multiple key_pairs, or only one key_pair in which case only the
+    /// associated records can be transferred.
+    pub sender_key_pairs: &'a Vec<UserKeyPair>,
+    pub asset: &'a AssetCode,
+    pub receivers: &'a [(UserPubKey, RecordAmount, bool)],
+    pub fee: RecordAmount,
+    pub bound_data: Vec<u8>,
+    pub xfr_size_requirement: Option<(usize, usize)>,
+}
+
+// (block_id, txn_id, [(uid, remember)])
+type CommittedTxn<'a> = (u64, u64, &'a mut [(u64, bool)]);
 
 // Inform the Transactions database that we have received memos for the given record UIDs. Return a list of
 // the transactions that are completed as a result.
@@ -275,6 +538,14 @@ async fn view_transaction<'a, L: Ledger, Meta: Serialize + DeserializeOwned + Se
         }
     }
     Ok(())
+}
+
+fn bigint_to_u256(i: BigInt) -> U256 {
+    let (sign, mut digits) = i.to_u64_digits();
+    assert_ne!(sign, Sign::Minus);
+    assert!(digits.len() <= 4);
+    digits.resize(4, 0);
+    U256(digits.try_into().unwrap())
 }
 
 fn u256_to_signed(u: U256) -> BigInt {
@@ -1548,7 +1819,7 @@ impl<'a, L: 'static + Ledger> LedgerState<'a, L> {
             }
         };
 
-        model.ledger_state_store.update_dynamic(self)?;
+        model.ledger_states.update_dynamic(self)?;
         Ok(summary)
     }
 
@@ -1752,7 +2023,7 @@ impl<'a, L: 'static + Ledger> LedgerState<'a, L> {
                     account.set_used().save()?;
                 }
             }
-            model.ledger_state_store.update_dynamic(self)?;
+            model.ledger_states.update_dynamic(self)?;
             Ok(definition)
         }
     }
@@ -1839,7 +2110,7 @@ impl<'a, L: 'static + Ledger> LedgerState<'a, L> {
             .with_description(description)
             .set_scan(scan)
             .save()?;
-        model.ledger_state_store.update_dynamic(self)?;
+        model.ledger_states.update_dynamic(self)?;
         // If we successfully updated our data structures, register the key with the
         // network. The storage transaction will revert if this fails.
         model.backend.register_user_key(&user_key).await?;
@@ -1876,7 +2147,7 @@ impl<'a, L: 'static + Ledger> LedgerState<'a, L> {
             .create(viewing_key.clone(), index)?
             .with_description(description)
             .save()?;
-        model.ledger_state_store.update_dynamic(self)?;
+        model.ledger_states.update_dynamic(self)?;
         Ok(viewing_key)
     }
 
@@ -1910,7 +2181,7 @@ impl<'a, L: 'static + Ledger> LedgerState<'a, L> {
             .create(freezing_key.clone(), index)?
             .with_description(description)
             .save()?;
-        model.ledger_state_store.update_dynamic(self)?;
+        model.ledger_states.update_dynamic(self)?;
 
         Ok(freezing_key)
     }
@@ -2070,7 +2341,7 @@ impl<'a, L: 'static + Ledger> LedgerState<'a, L> {
                 }
                 info.timeout = Some(timeout);
                 let stored_txn = model.transactions.create(uid, info)?;
-                model.ledger_state_store.update_dynamic(self)?;
+                model.ledger_states.update_dynamic(self)?;
                 stored_txn.clone()
             } else {
                 model.transactions.get(&TransactionUID::<L>(txn.hash()))?
@@ -2146,13 +2417,13 @@ impl<'a, L: Ledger> From<&LedgerState<'a, L>> for DynamicState<L> {
     }
 }
 
-/// Storage for the ledger state.
-pub struct LedgerStateStore<'a, L: Ledger> {
+/// Storage for the static and dynamic parts of the ledger state.
+pub struct LedgerStates<'a, L: Ledger> {
     static_store: RollingLog<EncryptingResourceAdapter<StaticState<'a>>>,
     dynamic_store: RollingLog<EncryptingResourceAdapter<DynamicState<L>>>,
 }
 
-impl<'a, L: 'static + Ledger> LedgerStateStore<'a, L> {
+impl<'a, L: 'static + Ledger> LedgerStates<'a, L> {
     /// Create a ledger state store.
     pub(crate) fn new(
         loader: &mut AtomicStoreLoader,
