+++ src/testing/tests.rs
@@ -1946,7 +1946,7 @@ pub mod generic_keystore_tests {
                 .await
                 .update(|KeystoreSharedState { model, .. }| async move {
                     let key = model
-                        .persistence
+                        .meta_store
                         .key_stream()
                         .derive_sub_tree("user".as_bytes())
                         .derive_user_key_pair(&model.sending_accounts.index().to_le_bytes());
